Extracted test runner into a seperate crate
