+++ crates/forge/tests/integration/common/runner.rs
@@ -163,7 +163,7 @@ macro_rules! test_case {
 #[macro_export]
 macro_rules! assert_passed {
     ($result:expr) => {{
-        use forge::test_case_summary::TestCaseSummary;
+        use test_runner::test_case_summary::TestCaseSummary;
         use $crate::integration::common::runner::TestCase;
 
         let result = TestCase::find_test_result(&$result);
@@ -184,7 +184,7 @@ macro_rules! assert_passed {
 #[macro_export]
 macro_rules! assert_failed {
     ($result:expr) => {{
-        use forge::test_case_summary::TestCaseSummary;
+        use test_runner::test_case_summary::TestCaseSummary;
 
         use $crate::integration::common::runner::TestCase;
 
@@ -206,7 +206,7 @@ macro_rules! assert_failed {
 #[macro_export]
 macro_rules! assert_case_output_contains {
     ($result:expr, $test_case_name:expr, $asserted_msg:expr) => {{
-        use forge::test_case_summary::TestCaseSummary;
+        use test_runner::test_case_summary::TestCaseSummary;
 
         use $crate::integration::common::runner::TestCase;
