+++ crates/test-collector/Cargo.toml
@@ -40,6 +40,7 @@ thiserror.workspace = true
 num-bigint.workspace = true
 smol_str = { version = "0.2.0", features = ["serde"] }
 num-integer = "0.1"
+once_cell = "1.18.0"
 
 
 [lib]
