+++ crates/test-runner/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
-name = "forge"
+name = "test-runner"
 version.workspace = true
 edition.workspace = true
+repository.workspace = true
+license.workspace = true
+license-file.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -16,7 +19,6 @@ starknet_api.workspace = true
 tempfile.workspace = true
 rayon.workspace = true
 test-collector = { path = "../test-collector" }
-test-runner = { path = "../test-runner" }
 cheatnet = { path = "../cheatnet" }
 conversions = { path = "../conversions" }
 cairo-lang-runner.workspace = true
@@ -59,10 +61,3 @@ toml_edit = "0.9.1"
 flatten-serde-json = "0.1.0"
 rand = "0.8.5"
 smol_str = "0.2.0"
-
-[lib]
-name = "forge"
-
-[[bin]]
-name = "snforge"
-path = "src/main.rs"
